+++ owncloud/appManagement.js
@@ -3,7 +3,7 @@
 ///////////////////////////////////
 
 var Promise = require('promise');
-var parser = require('xml-js');
+var parser = require('./xmlParser/');
 var helpers;
 
 /**
@@ -46,25 +46,15 @@ apps.prototype.getApps = function() {
     return new Promise((resolve, reject) => {
         Promise.all([allAppsP, allEnabledAppsP])
             .then(apps => {
-                var tree = parser.xml2js(apps[0].body, {
-                    compact: true
-                });
-                tree = helpers._cleanseJson(tree);
+                var tree = parser.xml2js(apps[0].body);
                 var statuscode = parseInt(helpers._checkOCSstatusCode(tree));
                 if (statuscode === 999) {
                     reject("Provisioning API has been disabled at your instance");
                     return;
                 }
 
-                var allApps = parser.xml2js(apps[0].body, {
-                    compact: true
-                });
-                allApps = helpers._cleanseJson(allApps).ocs.data.apps.element;
-
-                var allEnabledApps = parser.xml2js(apps[1].body, {
-                    compact: true
-                });
-                allEnabledApps = helpers._cleanseJson(allEnabledApps).ocs.data.apps.element;
+                var allApps = parser.xml2js(apps[0].body).ocs.data.apps.element;
+                var allEnabledApps = parser.xml2js(apps[1].body).ocs.data.apps.element;
 
                 for (var i = 0; i < allApps.length; i++) {
                     send[allApps[i]] = false;
@@ -100,10 +90,7 @@ apps.prototype.getAttribute = function(app, key) {
     return new Promise((resolve, reject) => {
         helpers._makeOCSrequest('GET', helpers.OCS_SERVICE_PRIVATEDATA, send)
             .then(data => {
-                var elements = parser.xml2js(data.body, {
-                    compact: true
-                });
-                elements = helpers._cleanseJson(elements).ocs.data.element;
+                var elements = parser.xml2js(data.body).ocs.data.element;
 
                 if (key) {
                     if (!elements) {
