+++ owncloud/shareManagement.js
@@ -3,7 +3,7 @@
 /////////////////////////////
 
 var Promise = require('promise');
-var parser = require('xml-js');
+var parser = require('./xmlParser/');
 var utf8 = require('utf8');
 var shareInfo = require('./shareInfo.js');
 var helpers;
@@ -71,11 +71,7 @@ shares.prototype.shareFileWithLink = function(path, optionalParams) {
         helpers._makeOCSrequest('POST', helpers.OCS_SERVICE_SHARE, 'shares', postData)
             .then(data => {
                 data.body = utf8.encode(data.body);
-                var shareDetails = parser.xml2js(data.body, {
-                    compact: true
-                });
-                shareDetails = helpers._cleanseJson(shareDetails);
-                shareDetails = shareDetails.ocs.data;
+                var shareDetails = parser.xml2js(data.body).ocs.data;
                 var share = new shareInfo(shareDetails);
 
                 resolve(share);
@@ -115,12 +111,9 @@ shares.prototype.shareFileWithUser = function(path, username, optionalParams) {
     return new Promise((resolve, reject) => {
         helpers._makeOCSrequest('POST', helpers.OCS_SERVICE_SHARE, 'shares', postData)
             .then(data => {
-                var shareData = parser.xml2js(data.body, {
-                    compact: true
-                });
-                shareData = helpers._cleanseJson(shareData).ocs.data;
-
+                var shareData = parser.xml2js(data.body).ocs.data
                 var share = new shareInfo(shareData);
+
                 resolve(share);
             }).catch(error => {
                 reject(error);
@@ -151,11 +144,9 @@ shares.prototype.shareFileWithGroup = function(path, groupName, optionalParams) 
     return new Promise((resolve, reject) => {
         helpers._makeOCSrequest('POST', helpers.OCS_SERVICE_SHARE, 'shares', postData)
             .then(data => {
-                var shareData = parser.xml2js(data.body, {
-                    compact: true
-                });
-                shareData = helpers._cleanseJson(shareData).ocs.data;
+                var shareData = parser.xml2js(data.body).ocs.data;
                 var share = new shareInfo(shareData);
+
                 resolve(share);
             }).catch(error => {
                 reject(error);
@@ -209,11 +200,7 @@ shares.prototype.getShares = function(path, optionalParams) {
     return new Promise((resolve, reject) => {
         helpers._makeOCSrequest('GET', helpers.OCS_SERVICE_SHARE, data)
             .then(data => {
-                var elements = parser.xml2js(data.body, {
-                    compact: true
-                });
-                elements = helpers._cleanseJson(elements).ocs.data.element || [];
-
+                var elements = parser.xml2js(data.body).ocs.data.element || [];
                 var shares = [];
 
                 if (elements && elements.constructor !== Array) {
@@ -265,10 +252,7 @@ shares.prototype.getShare = function(shareId) {
         }
         helpers._makeOCSrequest('GET', helpers.OCS_SERVICE_SHARE, 'shares/' + shareId.toString())
             .then(data => {
-                var shareData = parser.xml2js(data.body, {
-                    compact: true
-                });
-                shareData = helpers._cleanseJson(shareData).ocs.data.element;
+                var shareData = parser.xml2js(data.body).ocs.data.element;
                 var share = new shareInfo(shareData);
 
                 resolve(share);
@@ -287,10 +271,8 @@ shares.prototype.listOpenRemoteShare = function() {
     return new Promise((resolve, reject) => {
         helpers._makeOCSrequest('GET', helpers.OCS_SERVICE_SHARE, 'remote_shares/pending')
             .then(data => {
-                var shares = parser.xml2js(data.body, {
-                    compact: true
-                });
-                shares = helpers._cleanseJson(shares).ocs.data.element || [];
+                var shares = parser.xml2js(data.body).ocs.data.element || [];
+                
                 resolve(shares);
             }).catch(error => {
                 reject(error);
