+++ owncloud/xmlParser/index.js
@@ -1,0 +1,178 @@
+var parser = require('xml-js');
+var myParser = {};
+
+/**
+ * The main function
+ * @param  {string}  xml    xml which needs to be parsed
+ * @param  {object} [xmlns] contains all namespaces
+ * @return {object}         parsed js object
+ */
+myParser.xml2js = function (xml, xmlns) {
+    var parsed = parser.xml2js(xml, {
+        compact: true
+    });
+
+    if (xmlns) {
+        // Keep Namespace
+        parsed = myParser.keepNamespace(parsed, xmlns);
+    }
+
+    parsed = myParser.cleanseJson(parsed);
+    return parsed;
+};
+
+/**
+ * Keeps the namespace
+ * @param  {object} json In which to replace namespace
+ * @param  {object} ns   Namespace object
+ * @return {object}      Namespace replaced object
+ */
+myParser.keepNamespace = function(json, ns) {
+    var nsKeys = Object.keys(ns);
+
+    for (var key in json) {
+        var parseKey = parseKeyNS(key);
+        if (key.indexOf(':') > -1 && nsKeys.indexOf(parseKey) > -1) {
+            var index = nsKeys.indexOf(parseKey);
+            var prop = '{' + ns[nsKeys[index]] + '}' + key.split(':')[1];
+
+            json[prop] = json[key];
+            json[prop] = recursiveNS(json[prop], ns);
+        } else {
+            json[key] = recursiveNS(json[key], ns);
+        }
+    }
+
+    json = myParser.deleteDuplicates(json, ns);
+    return json;
+};
+
+/**
+ * Intermediate of keepNamespace()
+ */
+myParser.deleteDuplicates = function(json, ns) {
+    var ret = {};
+    var nsKeys = Object.keys(ns);
+    if (json.constructor === Array) {
+        ret = [];
+    }
+
+    if (typeof(json) !== 'object') {
+        return json;
+    }
+
+    for (var key in json) {
+        if (json.constructor === Array) {
+            ret.push(recursiveDeleteDuplicates(json[key], ns));
+        }
+        if (key.indexOf(':') > -1) {
+            var parseKey = parseKeyNS(key);
+            if (nsKeys.indexOf(parseKey) === -1) {
+                ret[key] = recursiveDeleteDuplicates(json[key], ns);
+            }
+        } else if (json.constructor !== Array) {
+            ret[key] = json[key];
+        }
+    }
+    return ret;
+};
+
+/**
+ * This function removes the "_text" attribute introduced by the XML parser
+ * For more info, check the NPM page of "xml-js", and see response format of the parser
+ * @param  {object} json object to cleanse
+ * @return {object}      cleaned object
+ */
+myParser.cleanseJson = function(json) {
+    for (var key in json) {
+        var a = recursiveCleanse(json[key]);
+        json[key] = a;
+    }
+    return json;
+};
+
+/**
+ * HELPER FOR keepNamespace()
+ */
+function recursiveNS(json, ns) {
+    if (typeof(json) !== 'object') {
+        return json;
+    }
+    var nsKeys = Object.keys(ns);
+
+    for (var key in json) {
+        var parseKey = parseKeyNS(key);
+        if (key.indexOf(':') > -1 && nsKeys.indexOf(parseKey) > -1) {
+            var index = nsKeys.indexOf(parseKey);
+            var prop = '{' + ns[nsKeys[index]] + '}' + key.split(':')[1];
+
+            json[prop] = json[key];
+            json[prop] = recursiveNS(json[prop], ns);
+        } else {
+            json[key] = recursiveNS(json[key], ns);
+        }
+    }
+    return json;
+}
+
+/**
+ * HELPER FOR deleteDuplicates()
+ */
+function recursiveDeleteDuplicates(json, ns) {
+    if (typeof(json) !== 'object') {
+        return json;
+    }
+
+    var nsKeys = Object.keys(ns);
+    var ret = {};
+    if (json.constructor === Array) {
+        ret = [];
+    }
+    for (var key in json) {
+        if (json.constructor === Array) {
+            ret.push(recursiveDeleteDuplicates(json[key], ns));
+        }
+        if (key.indexOf(':') > -1) {
+            var parseKey = parseKeyNS(key);
+            if (nsKeys.indexOf(parseKey) === -1) {
+                ret[key] = recursiveDeleteDuplicates(json[key], ns);
+            }
+        } else if (json.constructor !== Array) {
+            ret[key] = json[key];
+        }
+    }
+    return ret;
+}
+
+/**
+ * HELPER FOR cleanseJson()
+ */
+function recursiveCleanse(json) {
+    if (typeof(json) !== 'object') {
+        return json;
+    }
+
+    for (var key in json) {
+        if (key === '_text') {
+            return json[key];
+        }
+        json[key] = recursiveCleanse(json[key]);
+    }
+    return json;
+}
+
+/**
+ * parses a key from d: to d and {DAV:} to DAV
+ * @param  {string} key key to be parsed
+ * @return {string}     parsed key
+ */
+function parseKeyNS(key) {
+    var parseKey = key.split(':')[0];
+    if (parseKey.slice(0, 1) === '{') {
+        parseKey = parseKey.slice(1);
+    }
+
+    return parseKey;
+}
+
+module.exports = myParser;
